// src/app/Swagger/models/user.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { FinalReservation } from './finalReservation';
import { GrantedAuthority } from './grantedAuthority';
import { PreReservation } from './preReservation';
import { Review } from './review';

export interface User {
  id?: number;
  firstName?: string;
  lastName?: string;
  email?: string;
  password?: string;
  role?: User.RoleEnum;
  profilePicture?: string;
  preReservations?: Array<PreReservation>;
  finalReservations?: Array<FinalReservation>;
  reviews?: Array<Review>;
  enabled?: boolean;
  username?: string;
  authorities?: Array<GrantedAuthority>;
  accountNonExpired?: boolean;
  accountNonLocked?: boolean;
  credentialsNonExpired?: boolean;
  city?: string;
  dateOfBirth?: Date;
}
export namespace User {
  export type RoleEnum = 'PROFESIONAL' | 'PARTICULAR' | 'ADMIN';
  export const RoleEnum = {
    PROFESIONAL: 'PROFESIONAL' as RoleEnum,
    PARTICULAR: 'PARTICULAR' as RoleEnum,
    ADMIN: 'ADMIN' as RoleEnum,
  };
}